+++ include/steps/htseq_count.py
@@ -22,14 +22,11 @@ class HtSeqCount(AbstractStep):
 
         #        self.add_connection('in/alignments')
         # the BAM files
-        self.add_connection(
-            'in/alignments',
-            constraints = {'min_files_per_run': 1, 'max_files_per_run': 1}
-        )
+        self.add_connection('in/alignments')
 
         # the feature file provided by another step (e.g. cuffmerge)
         self.add_connection('in/features', optional=True, format='gtf',
-            description='reference assembly', constraints = {'total_files': 1}
+            description='reference assembly'
         )
 
         # the counts per alignment